refactor: move locations of jwt unit tests
